requirements.txt
chardet==5.2.0
et_xmlfile==2.0.0
lxml==5.4.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
PyPDF2==3.0.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.1
PyQt6-sip==13.10.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
reportlab==4.4.1
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
pyinstaller==6.13.0
